# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Open-source drug information gateway. Provides NDC lookup, therapeutic class search, drug interactions, RxNorm fuzzy search, and structured product label data — all through a clean REST API with Redis caching, per-key rate limiting, and circuit breaker protection.\n\n## Getting Started\n1. **Create an API key:** `POST /admin/keys` with `Authorization: Bearer {ADMIN_SECRET}` and body `{\"app_name\": \"my-app\", \"origins\": [\"*\"], \"rate_limit\": 1000}`\n2. **Search for a drug:** `GET /v1/drugs/autocomplete?q=lipit` with header `X-API-Key: {your_key}` — returns matching drug names\n3. **Look up details:** `GET /v1/drugs/class?name=atorvastatin` — returns therapeutic classes and brand names\n4. **Get interactions:** `GET /v1/drugs/info?name=warfarin` — returns FDA label interaction warnings\n5. **Check multi-drug interactions:** `POST /v1/drugs/interactions` with `{\"drugs\": [{\"name\": \"warfarin\"}, {\"name\": \"aspirin\"}]}`",
        "title": "drug-gate API",
        "contact": {
            "name": "drug-gate on GitHub",
            "url": "https://github.com/finish06/drug-gate"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/finish06/drug-gate/blob/main/LICENSE"
        },
        "version": "0.9.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/drugs/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Look up a single drug by name or NDC and return a consolidated info card with SPL metadata, drug interactions, contraindications, warnings, and adverse reactions. When an NDC is provided, it is normalized and resolved to a drug name internally. Provide either name or ndc; if both are given, ndc takes precedence.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spl"
                ],
                "summary": "Drug info card with interactions",
                "parameters": [
                    {
                        "type": "string",
                        "example": "warfarin",
                        "description": "Drug name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "00069-3150",
                        "description": "NDC code (any format)",
                        "name": "ndc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Missing name or ndc",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "NDC not found or no SPL data",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/interactions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Submit 2 to 10 drug identifiers (by name or NDC) and receive cross-referenced interaction warnings from FDA SPL labels. Each drug's Section 7 (Drug Interactions) is searched for mentions of the other submitted drugs. NDC identifiers are resolved to drug names automatically. Use this endpoint to check a patient's medication list for potential interactions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spl"
                ],
                "summary": "Check drug interactions",
                "parameters": [
                    {
                        "description": "Drug identifiers to check",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionCheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionCheckResponse"
                        }
                    },
                    "400": {
                        "description": "Too few/many drugs or invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/drugs/spls": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Searches Structured Product Labels (SPLs) by drug name and returns paginated metadata from DailyMed. Each result includes the SPL title, set ID, published date, and version. Use the set ID from results to fetch full SPL detail with parsed interaction sections.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spl"
                ],
                "summary": "Search SPL documents",
                "parameters": [
                    {
                        "type": "string",
                        "example": "metformin",
                        "description": "Drug name to search",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "Results per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Missing name parameter",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/spls/{setid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves full SPL metadata and parsed safety sections from the SPL XML document, including Drug Interactions (Section 7), Contraindications, Warnings, and Adverse Reactions. Use this endpoint when you have a set ID from the SPL search results and need the complete label content.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spl"
                ],
                "summary": "Get SPL detail with interactions",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2c6ca939-8494-4b3b-8930-81c6c09b0125",
                        "description": "SPL set ID (UUID format)",
                        "name": "setid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.SPLDetail"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "SPL not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_finish06_drug-gate_internal_model.DrugCheckResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "has_interactions": {
                    "type": "boolean",
                    "example": true
                },
                "input_name": {
                    "type": "string",
                    "example": "warfarin"
                },
                "input_type": {
                    "type": "string",
                    "example": "name"
                },
                "resolved_name": {
                    "type": "string",
                    "example": "WARFARIN SODIUM"
                },
                "spl_setid": {
                    "type": "string",
                    "example": "b76006e5-1005-4a64-97a1-7cc20c0e6a1a"
                }
            }
        },
...
        "github_com_finish06_drug-gate_internal_model.DrugIdentifier": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "warfarin"
                },
                "ndc": {
                    "type": "string",
                    "example": "00069-3150"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.DrugInfoResponse": {
            "type": "object",
            "properties": {
                "adverse_reactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                },
                "contraindications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                },
                "drug_name": {
                    "type": "string",
                    "example": "warfarin"
                },
                "input_type": {
                    "type": "string",
                    "example": "name"
                },
                "input_value": {
                    "type": "string",
                    "example": "warfarin"
                },
                "interactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                },
                "spl": {
                    "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.SPLSource"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                }
            }
        },
...
        "github_com_finish06_drug-gate_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "message": {
                    "type": "string",
                    "example": "No drug found for NDC 99999-9999"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.InteractionCheckRequest": {
            "type": "object",
            "properties": {
                "drugs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugIdentifier"
                    }
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.InteractionCheckResponse": {
            "type": "object",
            "properties": {
                "checked_pairs": {
                    "type": "integer",
                    "example": 3
                },
                "drugs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugCheckResult"
                    }
                },
                "found_interactions": {
                    "type": "integer",
                    "example": 2
                },
                "interactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionMatch"
                    }
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.InteractionMatch": {
            "type": "object",
            "properties": {
                "drug_a": {
                    "type": "string",
                    "example": "warfarin"
                },
                "drug_b": {
                    "type": "string",
                    "example": "aspirin"
                },
                "section_title": {
                    "type": "string",
                    "example": "Drug Interactions"
                },
                "source": {
                    "type": "string",
                    "example": "warfarin"
                },
                "spl_setid": {
                    "type": "string",
                    "example": "b76006e5-1005-4a64-97a1-7cc20c0e6a1a"
                },
                "text": {
                    "type": "string",
                    "example": "Aspirin can increase the anticoagulant effect of warfarin."
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.InteractionSection": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string",
                    "example": "Concomitant use of warfarin and aspirin increases the risk of bleeding."
                },
                "title": {
                    "type": "string",
                    "example": "Drug Interactions"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {
                    "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.Pagination"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 50
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 1234
                },
                "total_pages": {
                    "type": "integer",
                    "example": 25
                }
            }
        },
...
        "github_com_finish06_drug-gate_internal_model.SPLDetail": {
            "type": "object",
            "properties": {
                "adverse_reactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                },
                "contraindications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                },
                "interactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                },
                "published_date": {
                    "type": "string",
                    "example": "2024-03-15"
                },
                "setid": {
                    "type": "string",
                    "example": "b76006e5-1005-4a64-97a1-7cc20c0e6a1a"
                },
                "spl_version": {
                    "type": "integer",
                    "example": 12
                },
                "title": {
                    "type": "string",
                    "example": "WARFARIN SODIUM tablet"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.InteractionSection"
                    }
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.SPLSource": {
            "type": "object",
            "properties": {
                "published_date": {
                    "type": "string",
                    "example": "2024-03-15"
                },
                "setid": {
                    "type": "string",
                    "example": "b76006e5-1005-4a64-97a1-7cc20c0e6a1a"
                },
                "spl_version": {
                    "type": "integer",
                    "example": 12
                },
                "title": {
                    "type": "string",
                    "example": "WARFARIN SODIUM tablet"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token set via `ADMIN_SECRET` environment variable. Format: `Bearer {secret}`. Required for all `/admin/*` endpoints.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Publishable API key for frontend applications. Create one via `POST /admin/keys`. Include in all `/v1/*` requests.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    },
...
}